requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
PyJWT==2.4.0
python-http-client==3.3.7
Rx==1.6.1
sendgrid==6.9.7
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
text-unidecode==1.3
